Break filter bar and title bar into separate components for clarity. Memoise the filter bar so it doesn't re-render as the count comes in from the new feature.
