+++ src/packages/admin-ui-components/src/title-bar/index.ts
@@ -1,0 +1,1 @@
+export * from './component';
